Monster commit. Details below.
Changed top to ps. Following reasons:
- ps is independent on locale (top turns e.g. CPU 2.3% to 2,3% depending on locale)
- ps does not produce the header that has to be shaved of
- ps is (arguably) more adaptable/configurable

Rewrote extract_processes and its test accordingly, so it works with the new ps solution.

Renamed get_slurm_project to get_slurm_projects (with s) as it will get all user-to-project mappings on the node.

Implemented get_slurm_projects. It will only get the active users (and their projects) on the current cluster, since the snap script will run on each node and does not need to know about other nodes.

Managed ignored users in another way that should (in my opinion) work better in future with config files.

I split take_snapshot into two functions, one to gather all data and return a list of lists, and another function that writes everything to file or stdout. This way, the first function is more easily testable and the second function could easily be changed to a database or whatever without tinkering with the first. Naming of the two functions may be unfortunate/unclear.

The second function uses a special wrapper function that enables easy writing to stdout or file.

Memory is saved in MiB, since the percentage is dependent on the node.

Slurm-specific stuff was tested on Stallo :)